Add find, height, and various iterators
Find and height are implemented together in the Node class as _find. The
Tree methods find and height return different their respective results
from _find. Iterators for in-order, pre-order, and post-order traversal
of the Tree were implemented, as well as level_order which traverses
only the tree Nodes at a specified height.