Add caesar.py and test_caesar.py
Skipped the original Caesar exercise from the Ruby section since I
accidentally worked ahead on the JS exercises and had solved it recently
in that language. The Test Driven Development section suggests writing
tests for a student's existing version of this project so I'm going to
write tests then do the exercise in Python to practice TDD more
directly. The tests are borrowed from the tests for this exercise in the
Odin Project javascript exercises repo.